Move report commands under report category, rename summarize to summary
