+++ src/cli/main.ts
@@ -56,18 +56,6 @@ import {
   loadTaskDefinitionsForCheck,
   runConfigCheck,
 } from "./commands/check.js";
-import {
-  createReportSummaryArtifact,
-  formatPrComment,
-  formatReportAggregate,
-  formatReportDiff,
-  formatReportSummary,
-  loadReportSummaryArtifactsFromDir,
-  parseReportSummary,
-  runReportAggregate,
-  runReportDiff,
-  runReportSummarize,
-} from "./commands/report.js";
 import { DuckDBStore } from "./storage/duckdb.js";
 import { createRunner } from "./runners/index.js";
 import { resolveTestIdentity } from "./identity.js";
@@ -140,12 +128,11 @@ export function createProgram(): Command {
   registerExecCommands(program);
   registerCollectCommands(program);
   registerImportCommands(program);
+  registerReportCommands(program);
   registerAnalyzeCommands(program);
   registerDebugCommands(program);
   registerPolicyCommands(program);
   registerDevCommands(program);
-  // registerReportCommands deferred:
-  // existing flat command with that name must be removed first (Task 6)
 
 async function collectKnownQuarantineTaskIds(
   cwd: string,
@@ -484,141 +471,6 @@ program
     process.exit(report.errors.length > 0 ? 1 : 0);
   });
 
-// --- report ---
-const reportCommand = program
-  .command("report")
-  .description("Summarize and diff normalized test reports")
-  .action(() => {
-    reportCommand.outputHelp();
-  });
-
-reportCommand
-  .command("summarize")
-  .description("Summarize a raw adapter report")
-  .requiredOption("--adapter <type>", "Adapter type (playwright, junit, vrt-migration, vrt-bench)")
-  .requiredOption("--input <file>", "Raw adapter report file")
-  .option("--bundle", "Wrap summary with shard metadata for aggregation")
-  .option("--shard <name>", "Shard name")
-  .option("--module <name>", "Module name")
-  .option("--offset <n>", "Shard offset")
-  .option("--limit <n>", "Shard limit")
-  .option("--matrix <pairs>", "Comma-separated matrix metadata (key=value)")
-  .option("--variant <pairs>", "Comma-separated variant metadata (key=value)")
-  .option("--meta <pairs>", "Comma-separated extra metadata (key=value)")
-  .option("--json", "Output JSON report")
-  .option("--markdown", "Output Markdown report")
-  .option("--pr-comment", "Output compact Markdown for PR comments")
-  .action(
-    (opts: {
-      adapter: string;
-      input: string;
-      bundle?: boolean;
-      shard?: string;
-      module?: string;
-      offset?: string;
-      limit?: string;
-      matrix?: string;
-      variant?: string;
-      meta?: string;
-      json?: boolean;
-      markdown?: boolean;
-      prComment?: boolean;
-    }) => {
-      const formatCount = [opts.json, opts.markdown, opts.prComment].filter(Boolean).length;
-      if (formatCount > 1) {
-        console.error("Error: choose one of --json, --markdown, or --pr-comment");
-        process.exit(1);
-      }
-      if (opts.bundle && opts.markdown) {
-        console.error("Error: --bundle cannot be combined with --markdown");
-        process.exit(1);
-      }
-
-      const summary = runReportSummarize({
-        adapter: opts.adapter,
-        input: readFileSync(resolve(opts.input), "utf-8"),
-      });
-      if (opts.bundle) {
-        console.log(
-          JSON.stringify(
-            createReportSummaryArtifact(summary, {
-              shard: opts.shard,
-              module: opts.module,
-              offset: opts.offset ? Number(opts.offset) : undefined,
-              limit: opts.limit ? Number(opts.limit) : undefined,
-              matrix: parseKeyValuePairs(opts.matrix),
-              variant: parseKeyValuePairs(opts.variant),
-              extra: parseKeyValuePairs(opts.meta),
-            }),
-            null,
-            2,
-          ),
-        );
-        return;
-      }
-      if (opts.prComment) {
-        console.log(formatPrComment(summary));
-        return;
-      }
-      console.log(
-        formatReportSummary(summary, opts.json ? "json" : "markdown"),
-      );
-    },
-  );
-
-reportCommand
-  .command("diff")
-  .description("Diff two normalized summaries or raw adapter reports")
-  .requiredOption("--base <file>", "Base summary or raw report file")
-  .requiredOption("--head <file>", "Head summary or raw report file")
-  .option("--adapter <type>", "Adapter type when diffing raw reports")
-  .option("--json", "Output JSON report")
-  .option("--markdown", "Output Markdown report")
-  .action(
-    (opts: {
-      base: string;
-      head: string;
-      adapter?: string;
-      json?: boolean;
-      markdown?: boolean;
-    }) => {
-      if (opts.json && opts.markdown) {
-        console.error("Error: choose either --json or --markdown");
-        process.exit(1);
-      }
-
-      const baseInput = readFileSync(resolve(opts.base), "utf-8");
-      const headInput = readFileSync(resolve(opts.head), "utf-8");
-      const base = opts.adapter
-        ? runReportSummarize({ adapter: opts.adapter, input: baseInput })
-        : parseReportSummary(baseInput);
-      const head = opts.adapter
-        ? runReportSummarize({ adapter: opts.adapter, input: headInput })
-        : parseReportSummary(headInput);
-      const diff = runReportDiff({ base, head });
-
-      console.log(formatReportDiff(diff, opts.json ? "json" : "markdown"));
-    },
-  );
-
-reportCommand
-  .command("aggregate <dir>")
-  .description("Aggregate shard-aware summary artifacts")
-  .option("--json", "Output JSON report")
-  .option("--markdown", "Output Markdown report")
-  .action((dir: string, opts: { json?: boolean; markdown?: boolean }) => {
-    if (opts.json && opts.markdown) {
-      console.error("Error: choose either --json or --markdown");
-      process.exit(1);
-    }
-
-    const aggregate = runReportAggregate({
-      summaries: loadReportSummaryArtifactsFromDir(resolve(dir)),
-    });
-    console.log(
-      formatReportAggregate(aggregate, opts.json ? "json" : "markdown"),
-    );
-  });
 
 // --- query ---
 program
